pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lumiverb"
version = "0.1.0"
description = "AI-powered photo and video library management"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}

# Base dependencies: API server + shared core (what the VPS needs)
dependencies = [
    # API & web framework
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # Database
    "sqlmodel>=0.0.14",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3.0",
    # Auth
    "pyjwt>=2.8",
    "bcrypt>=4.0",
    # Search
    "PyYAML>=6.0",
    # Shared utilities
    "python-ulid>=2.0",
    "requests>=2.25",
    # Image normalization (proxy → WebP, thumbnail generation)
    "pillow>=11.1",
]

[project.scripts]
lumiverb = "src.client.cli:main"

[project.optional-dependencies]
# CLI: local agent that talks to the API
cli = [
    "typer>=0.12",
    "rich>=13.0",
    "httpx>=0.27",
]
# Workers: image/video processing and AI (runs on the client machine)
workers = [
    "rawpy>=0.26.1",
    "pyvips>=3.1.1",
    "imagehash>=4.3.0",
    "opencv-python-headless>=4.8.0",
    "pyexiftool>=0.5",
    "torch>=2.10.0",
    "open-clip-torch>=3.3.0",
    "insightface>=0.7.3",
    "onnxruntime>=1.17.0",
    "faster-whisper>=1.0.0",
]
# All extras: CLI + workers (for local development and client installs)
all = [
    "lumiverb[cli]",
    "lumiverb[workers]",
]
dev = [
    "lumiverb[all]",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-order>=1.0",
    "testcontainers[postgres]>=4.0",
    "mypy>=1.8",
    "ruff>=0.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# =============================================================================
# pytest
# =============================================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "fast: no DB, no AI — pure unit tests, mocks, tmp_path only",
    "slow: requires DB (testcontainers Postgres)",
    "ai: requires real AI inference (Moondream Station running)",
    "quickwit: requires dev Quickwit at port 7281",
    "migration: Alembic migration tests (run separately)",
]
# Default: run fast and slow tests; exclude AI and migration suites
addopts = ["-m", "not migration and not ai and not quickwit"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# =============================================================================
# Ruff (linter + formatter)
# =============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter

# =============================================================================
# mypy
# =============================================================================
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
